refactor touch action of cell wrapper
