+++ lib/components/CellWrapper.tsx
@@ -1,10 +1,9 @@
 import React, { FC } from "react";
 import { useCellContext } from "./CellContext";
 import HiddenFocusTarget from "./HiddenFocusTarget";
+import { ColumnResize } from "./ColumnResize";
 import { useReactGridStore } from "../utils/reactGridStore";
 import { useReactGridId } from "./ReactGridIdProvider";
-import { useTheme } from "../hooks/useTheme";
-import { ResizeColumnBehavior } from "../behaviors/ResizeColumnBehavior";
 
 type CellWrapperProps = React.HTMLAttributes<HTMLDivElement> & {
   targetInputRef?: React.RefObject<HTMLInputElement | HTMLElement>;
@@ -14,25 +13,15 @@ type CellWrapperProps = React.HTMLAttributes<HTMLDivElement> & {
 const CellWrapper: FC<CellWrapperProps> = ({ children, targetInputRef, ...wrapperDivAttributes }) => {
   const { className: customClassName, style: customStyle } = wrapperDivAttributes;
   const ctx = useCellContext();
-  const theme = useTheme();
-
   const id = useReactGridId();
 
-  const resizingColId = useReactGridStore(id, (store) => store.resizingColId);
-  const columns = useReactGridStore(id, (store) => store.columns);
-  const shadowSize = useReactGridStore(id, (store) => store.shadowSize);
+  const enableColumnSelectionOnFirstRow = useReactGridStore(id, (store) => store.enableColumnSelectionOnFirstRow);
+  const enableRowSelectionOnFirstColumn = useReactGridStore(id, (store) => store.enableRowSelectionOnFirstColumn);
 
-  const onResizeColumn = useReactGridStore(id, (store) => store.onResizeColumn);
-  const setCurrentBehavior = useReactGridStore(id, (store) => store.setCurrentBehavior);
-  const setResizingColId = useReactGridStore(id, (store) => store.setResizingColId);
-
-  let shouldEnableColumnResize;
-
-  if (ctx.realRowIndex === 0) {
-    const cellColumn = columns.find((col) => col.id === ctx.colId);
-
-    shouldEnableColumnResize = resizingColId === undefined && onResizeColumn && cellColumn?.resizable && !shadowSize;
-  }
+  const disableTouchAction =
+    ctx.isFocused ||
+    (ctx.realRowIndex === 0 && enableColumnSelectionOnFirstRow) || // if column selection is enabled on first row
+    (ctx.realColumnIndex === 0 && enableRowSelectionOnFirstColumn); // if row selection is enabled on first column
 
   return (
     <div
@@ -44,27 +33,12 @@ const CellWrapper: FC<CellWrapperProps> = ({ children, targetInputRef, ...wrappe
         padding: ".2rem",
         textAlign: "center",
         position: "relative",
-        touchAction: "none",
+        touchAction: disableTouchAction ? "none" : "auto",
         ...customStyle,
         ...ctx.containerStyle,
       }}
     >
-      {shouldEnableColumnResize && (
-        <div
-          className="rg-resize-column"
-          onPointerDown={() => {
-            setResizingColId(ctx.colId);
-            setCurrentBehavior(ResizeColumnBehavior);
-          }}
-          css={{
-            cursor: "col-resize",
-            ...theme.resizeColumn.default,
-            "&:hover": {
-              ...theme.resizeColumn.hover,
-            },
-          }}
-        />
-      )}
+      {ctx.realRowIndex === 0 && <ColumnResize />}
       {children}
       {ctx.isFocused && <HiddenFocusTarget />}
     </div>
